Fix a few more MSVC level 3 warnings
 cbor.h(278): warning C4800: 'const uint16_t': forcing value to bool 'true' or 'false' (performance warning)
 cborencoder.c(295): warning C4334: '<<': result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift
 cborparser.c(677): warning C4800: 'uintptr_t': forcing value to bool 'true' or 'false' (performance warning)

+++ src/cborencoder.c
@@ -131,7 +131,7 @@ static inline CborError encode_number_no_update(CborEncoder *encoder, uint64_t u
             ++more;
         if (ui > 0xffffffffU)
             ++more;
-        bufstart -= 1 << more;
+        bufstart -= (size_t)1 << more;
         *bufstart = shiftedMajorType + Value8Bit + more;
     }
